requirements.txt
annotated-types==0.7.0
anyio==4.10.0
babel==2.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
courlan==1.3.2
dateparser==1.2.2
distro==1.9.0
faiss-cpu==1.12.0
fastapi==0.116.1
feedparser==6.0.11
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.9
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jusText==3.0.2
kafka-python==2.0.2
langdetect==1.0.9
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
neo4j==5.28.2
networkx==3.5
numpy==2.3.2
openai==1.104.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tld==0.13.1
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
